+++ scripts/release.sh
@@ -22,9 +22,6 @@ if ! grep -q "$version" "$root_dir/CHANGELOG.md"; then
     exit 1
 fi
 
-api_version=$(grep '## Interface:' "$toc_file" | grep -oP '\d+')
-game_version=$(curl -H "X-Api-Token: $CURSEFORGE_API_TOKEN" "https://wow.curseforge.com/api/game/versions" 2>/dev/null | jq -r ".[] | select(.apiVersion == \"$api_version\").id")
-
 zip_file="$build_dir/$addon_name-$version.zip"
 
 if [ -e "$zip_file" ]; then
@@ -62,11 +59,13 @@ done < CHANGELOG.md
 changelog=$(sed -e '/[^[:space:]]/,$!d' -e :a -e '/^[[:space:]]*$/{$d;N;ba' -e '}' "$changelog_file")
 echo "$changelog" > "$changelog_file"
 
+game_versions="9641,9894,10272"
+
 # CURSEFORGE
 metadata=$(jq -n \
                 --arg changelog "$changelog" \
-                --arg gameVersion "$game_version" \
-                '{changelog: $changelog, changelogType: "markdown", gameVersions: [$gameVersion | tonumber], releaseType: "beta"}')
+                --arg game_versions "$game_versions" \
+                '{changelog: $changelog, changelogType: "markdown", gameVersions: $game_versions | split(",") | map(tonumber), releaseType: "beta"}')
 
 curl --http1.1 \
     -H "X-Api-Token: $CURSEFORGE_API_TOKEN" \
@@ -80,7 +79,7 @@ curl --http1.1 \
     -F "id=26606" \
     -F "version=$version" \
     -F "changelog=$changelog" \
-    -F "compatible=$game_version" \
+    -F "compatible=$game_versions" \
     -F "updatefile=@$zip_file" \
     https://api.wowinterface.com/addons/update 1>&2